% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_generator}
\alias{extract_generator}
\title{Master function to extract features}
\usage{
extract_generator(feature)
}
\arguments{
\item{feature}{Feature of interest}
}
\value{
Function of interest
}
\description{
Master function to extract features
}
