% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Extract metadata from a particular package}
\usage{
extract_metadata(id)
}
\arguments{
\item{id}{A package ID that can be extracted from the function \code{\link{lpgc_search}}.}
}
\value{
A \code{tibble} with the following fields: author, maintainer, id, name, notes and creation date
}
\description{
Extract metadata from a particular package
}
\examples{
\dontrun{df <- extract_metadata(id = "47ea65a4-bc0e-42e1-8bfc-8bd3508642e6")}
}
