% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpgc_categories.R
\name{lpgc_categories}
\alias{lpgc_categories}
\title{Datasets available for each category}
\usage{
lpgc_categories(category)
}
\arguments{
\item{category}{Category of interest according to \code{lpgc_show_categories}}
}
\value{
A \code{\link[tibble]{tibble}} with author, maintainer, id, name, extra notes and creation date.
}
\description{
Datasets available for each category
}
\examples{
\dontrun{df <- lpgc_categories("salud")}
}
\seealso{
\code{\link{lpgc_show_categories}}
}
