% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpgc_load.R
\name{lpgc_load}
\alias{lpgc_load}
\title{Load the data from Las Palmas de Gran Canaria City Council.}
\usage{
lpgc_load(id)
}
\arguments{
\item{id}{A package ID that can be extracted from the function \code{\link{lpgc_search}}.}
}
\value{
A list with two slots:
\itemize{
 \item A \code{\link[tibble]{tibble}} with metadata provided by the site.
 \item If a single file, a \code{\link[tibble]{tibble}} with the data. If multiple files, a list with as many slots as files are read.
 }
}
\description{
Load the data from Las Palmas de Gran Canaria City Council.
}
\examples{
\dontrun{lpgc_load(id = "47ea65a4-bc0e-42e1-8bfc-8bd3508642e6")}
}
