% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpgc_search.R
\name{lpgc_search}
\alias{lpgc_search}
\title{Search datasets by keyword in the open data site of the city of Las Palmas de Gran Canaria}
\usage{
lpgc_search(keywords)
}
\arguments{
\item{keywords}{Words that describe the data that is intended to be found (in Spanish). Spaces may be included and there is no distinction between uppercase or lowercase.}
}
\value{
A \code{\link[tibble]{tibble}} with:
\itemize{
 \item Author
 \item Maintainer
 \item ID
 \item Name
 \item Notes
 \item Creation date
 }
}
\description{
Search datasets by keyword in the open data site of the city of Las Palmas de Gran Canaria
}
\examples{
\dontrun{lpgc_search(keywords = "medio ambiente")}
}
\seealso{
\code{\link{lpgc_load}}
}
