% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_quality_datasets}
\alias{test_quality_datasets}
\title{Function to test if datasets have the proper quality}
\usage{
test_quality_datasets(test_type)
}
\arguments{
\item{test_type}{Should be 'empty' for testing if data slot is empty and 'read' for testing if each dataset can be read.}
}
\value{
Proportion of datasets passing the specified test
}
\description{
Function to test if datasets have the proper quality
}
