% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{lpgc_show_categories}
\alias{lpgc_show_categories}
\title{Categories of open data available on the City Council of Las Palmas de Gran Canaria.}
\format{An object of class \code{character} of length 22.}
\usage{
lpgc_show_categories
}
\description{
Categories of open data available on the City Council of Las Palmas de Gran Canaria.
}
\examples{
openlpgc_categories
}
\keyword{datasets}
